move v2.events to v2.core.events and add some context tests
